flags

+++ srcs/process_loop.c
@@ -6,7 +6,7 @@
 /*   By: amarzial <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/04 14:56:51 by amarzial          #+#    #+#             */
-/*   Updated: 2017/03/08 16:01:28 by amarzial         ###   ########.fr       */
+/*   Updated: 2017/03/09 11:20:49 by amarzial         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ void	run_cycle(t_vm *vm)
 void	vm_loop(t_vm *vm, t_options *opt)
 {
 	while (vm->process_count && \
-	(!opt->dump || (vm->total_cycles < opt->dump_cycles)))
+	(!opt->dump || (vm->total_cycles < (unsigned int)opt->dump_cycles)))
 	{
 		vm->cycle++;
 		vm->total_cycles++;
